cli: util: split sub commands into modules
